Moved postorder_traversal
