+++ sympy/core/tests/test_traversal.py
@@ -1,9 +1,9 @@
 from sympy.core.basic import Basic
 from sympy.core.compatibility import default_sort_key
 from sympy.core.symbol import symbols
-from sympy.core.traversal import preorder_traversal, use
-from sympy import expand, factor, I
-
+from sympy.core.traversal import preorder_traversal, use, postorder_traversal
+from sympy import expand, factor, I, Piecewise, S, Integral, Tuple
+from sympy.functions.elementary.piecewise import ExprCondPair
 
 b1 = Basic()
 b2 = Basic(b1)
@@ -53,3 +53,29 @@ def test_use():
     assert use(f, factor, level=1, kwargs=kwargs) == (x + I)**2*(x - I)**2 - 1
     assert use(f, factor, level=2, kwargs=kwargs) == (x + I)**2*(x - I)**2 - 1
     assert use(f, factor, level=3, kwargs=kwargs) == (x**2 + 1)**2 - 1
+
+
+def test_postorder_traversal():
+    x, y, z, w = symbols('x y z w')
+    expr = z + w*(x + y)
+    expected = [z, w, x, y, x + y, w*(x + y), w*(x + y) + z]
+    assert list(postorder_traversal(expr, keys=default_sort_key)) == expected
+    assert list(postorder_traversal(expr, keys=True)) == expected
+
+    expr = Piecewise((x, x < 1), (x**2, True))
+    expected = [
+        x, 1, x, x < 1, ExprCondPair(x, x < 1),
+        2, x, x**2, S.true,
+        ExprCondPair(x**2, True), Piecewise((x, x < 1), (x**2, True))
+    ]
+    assert list(postorder_traversal(expr, keys=default_sort_key)) == expected
+    assert list(postorder_traversal(
+        [expr], keys=default_sort_key)) == expected + [[expr]]
+
+    assert list(postorder_traversal(Integral(x**2, (x, 0, 1)),
+        keys=default_sort_key)) == [
+            2, x, x**2, 0, 1, x, Tuple(x, 0, 1),
+            Integral(x**2, Tuple(x, 0, 1))
+        ]
+    assert list(postorder_traversal(('abc', ('d', 'ef')))) == [
+        'abc', 'd', 'ef', ('d', 'ef'), ('abc', ('d', 'ef'))]
