+++ sympy/utilities/__init__.py
@@ -3,7 +3,7 @@ SymPy.
 """
 from .iterables import (flatten, group, take, subsets,
     variations, numbered_symbols, cartes, capture, dict_merge,
-    postorder_traversal, interactive_traversal,
+    interactive_traversal,
     prefixes, postfixes, sift, topological_sort, unflatten,
     has_dups, has_variety, reshape, default_sort_key, ordered,
     rotations)
@@ -20,7 +20,7 @@ from .timeutils import timed
 
 __all__ = [
     'flatten', 'group', 'take', 'subsets', 'variations', 'numbered_symbols',
-    'cartes', 'capture', 'dict_merge', 'postorder_traversal',
+    'cartes', 'capture', 'dict_merge',
     'interactive_traversal', 'prefixes', 'postfixes', 'sift',
     'topological_sort', 'unflatten', 'has_dups', 'has_variety', 'reshape',
     'default_sort_key', 'ordered', 'rotations',
